handle unknown block number errors more gracefully
"Unknown block number" generally happens only when the block number
requested has not yet been synced to the node that is being called,
but is expected it will be soon.

This adds retrying logic if "Unknown block number" errors are hit and
retries until a request timeout limit is hit, which should generally
give the nodes enough time to sync up.